Moved Basic tests into own namespace
