+++ grate.unittests/Basic/Infrastructure/GrateEnvironment_.cs
@@ -3,7 +3,7 @@ using FluentAssertions;
 using grate.Infrastructure;
 using NUnit.Framework;
 
-namespace grate.unittests.Infrastructure
+namespace grate.unittests.Basic.Infrastructure
 {
     [TestFixture]
     [Category("Basic")]
