+++ grate.unittests/grate.unittests.csproj
@@ -29,10 +29,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Folder Include="Infrastructure\SqlServer" />
-  </ItemGroup>
-
-  <ItemGroup>
     <Compile Remove="Oracle\**" />
   </ItemGroup>
   <ItemGroup>
